requirements.txt
fastapi
uvicorn[standard]
httpx
sqlmodel
psycopg[binary]
alembic
APScheduler
python-dotenv
python-multipart
